Add unique file for the habit details calendar code
